Refactor bot config validator
I've moved the validator back into the BotConfig class. It seems better
to have it near the fields it validates. It's a little weird still,
because it's registered as a validator for token but also validates
token_file, but there doesn't seem to be any way to register class-level
validators.